+++ src/main/java/no/api/freemarker/java8/time/LocalDateAdapter.java
@@ -27,9 +27,7 @@ import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.util.List;
 
-import static no.api.freemarker.java8.time.DateTimeTools.METHOD_FORMAT;
-import static no.api.freemarker.java8.time.DateTimeTools.METHOD_UNKNOWN_MSG;
-import static no.api.freemarker.java8.time.DateTimeTools.createDateTimeFormatter;
+import static no.api.freemarker.java8.time.DateTimeTools.*;
 
 /**
  * LocalDateAdapter adds basic format support for {@link LocalDate} too FreeMarker 2.3.23 and above.
@@ -45,6 +43,8 @@ public class LocalDateAdapter extends AbstractAdapter<LocalDate> implements Adap
     public TemplateModel get(String s) throws TemplateModelException {
         if (METHOD_FORMAT.equals(s)) {
             return new LocalDateFormatter(getObject());
+        } else if(METHOD_EQUALS.equals(s) || METHOD_AFTER.equals(s) || METHOD_BEFORE.equals(s)) {
+            return new LocalDateChecker(getObject(), s);
         }
         throw new TemplateModelException(METHOD_UNKNOWN_MSG + s);
     }
@@ -58,6 +58,30 @@ public class LocalDateAdapter extends AbstractAdapter<LocalDate> implements Adap
         @Override
         public Object exec(List list) throws TemplateModelException {
             return getObject().format(createDateTimeFormatter(list, 0, DateTimeFormatter.ISO_LOCAL_DATE));
+        }
+    }
+
+    public class LocalDateChecker extends AbstractChecker<LocalDate> implements TemplateMethodModelEx {
+        private String method;
+
+        public LocalDateChecker(LocalDate obj, String method) {
+            super(obj);
+            this.method = method;
+        }
+
+        @SuppressWarnings("Duplicates")
+        @Override
+        public Object exec(List list) throws TemplateModelException {
+            LocalDateAdapter adapter = (LocalDateAdapter) list.get(0);
+            switch(method) {
+                case METHOD_EQUALS:
+                    return getObject().isEqual(adapter.getObject());
+                case METHOD_AFTER:
+                    return getObject().isAfter(adapter.getObject());
+                case METHOD_BEFORE:
+                    return getObject().isBefore(adapter.getObject());
+            }
+            throw new TemplateModelException("method not implemented");
         }
     }
 }